Add files via upload
The way this program interacts with the user is the following.
It asks you for input about the sorting method that you want to use. It generates the tests itself based on <h2> srand(time(NULL)) </h2> to generate the most random possible tests based on time.
It adds more functionality by showing to the user the time and sorting method in each separate file that it generates for each test.